Korriger navngiving av modeller
