+++ app/src/main/kotlin/no/nav/aap/arenaoppslag/modeller/Periode.kt
@@ -16,17 +16,6 @@ data class Periode(
     }
 }
 
-data class VedtakFakta(
-    var dagsmbt: Int,
-    var dags: Int,
-    var barntill: Int,
-    var satsbarntg: Int,
-    var barnmston: Int,
-    var dagsfsam: Int,
-    var justertg: String?,
-)
-
-
 data class PeriodeMed11_17(
     val periode: Periode,
     val aktivitetsfaseKode: String,
