+++ app/src/main/kotlin/no/nav/aap/arenaoppslag/modeller/VedtakFakta.kt
@@ -1,0 +1,11 @@
+package no.nav.aap.arenaoppslag.modeller
+
+data class VedtakFakta(
+    var dagsmbt: Int,
+    var dags: Int,
+    var barntill: Int,
+    var satsbarntg: Int,
+    var barnmston: Int,
+    var dagsfsam: Int,
+    var justertg: String?,
+)
